+++ sql/v006to007.sql
@@ -1,0 +1,3 @@
+ALTER TABLE `location` ADD `rank` INT NULL AFTER `id`;
+ALTER TABLE `location` ADD INDEX (`rank`);
+ALTER TABLE `location` ADD `gr_color` VARCHAR(6) NULL AFTER `on_graph`;